kafka_python/requirements.txt.txt
confluent-kafka==2.4.0
kafka-python==2.0.2
kafka-python-ng==2.2.2